Create data directory in /src and move partnersAndSponsorsLogos file inside
