--- src/components/grid-aware/PartnersAndSponsorsBlock/stories/partnersAndSponsorsLogos.js
@@ -1,95 +1,0 @@
-import benetechLogo from "./benetech-logo.png";
-import bridgeLogo from "./bridge-logo.png";
-import ciscoLogo from "./cisco-logo.png";
-import codeForSFLogo from "./code-for-sf-logo.png";
-import codeTenderloinLogo from "./code-tenderloin-logo.png";
-import compassFamilyServicesLogo from "./compass-family-services-logo.png";
-import evictionDefenseLogo from "./eviction-defense-logo.png";
-import gsmaLogo from "./gsma-logo.png";
-import homeownershipSFLogo from "./homeownership-sf-logo.png";
-import justiceAndDiversityCenterLogo from "./justice-and-diversity-center-logo.png";
-import larkinStreetLogo from "./larkin-street-logo.png";
-import mohcdLogo from "./mohcd-logo.png";
-import monkeybrainsLogo from "./monkeybrains-logo.png";
-import pagerdutyLogo from "./pagerduty-logo.png";
-import pcsLogo from "./pcs-logo.png";
-import sfFamiliesLogo from "./sf-families-logo.png";
-import uberLogo from "./uber-logo.png";
-import weworkLogo from "./wework-logo.png";
-
-export default [
-  {
-    url: mohcdLogo,
-    alt:
-      "Logo of Mayor's Office of Housing and Community Development organization.",
-  },
-  {
-    url: justiceAndDiversityCenterLogo,
-    alt:
-      "Logo of Justice and Diversity Center of The Bar Association of San Francisco organization.",
-  },
-  {
-    url: sfFamiliesLogo,
-    alt: "Logo of SF Families online public service directory organization.",
-  },
-  {
-    url: compassFamilyServicesLogo,
-    alt: "Logo of Compass Family Services.",
-  },
-  {
-    url: larkinStreetLogo,
-    alt: "Logo of Larkin Street Youth Services nonprofit organization.",
-  },
-  {
-    url: evictionDefenseLogo,
-    alt: "Logo of Eviction Defense Collaborative.",
-  },
-  {
-    url: homeownershipSFLogo,
-    alt: "Logo of Homeownership education and opportunity",
-  },
-  {
-    url: benetechLogo,
-    alt: "Logo of Benetech nonprofit organization.",
-  },
-  {
-    url: ciscoLogo,
-    alt: "Logo of Cisco Systems networking hardware company.",
-  },
-  {
-    url: codeTenderloinLogo,
-    alt: "Logo of Code Tenderloin.",
-  },
-  {
-    url: monkeybrainsLogo,
-    alt: "Logo of monkeybrains.",
-  },
-  {
-    url: pagerdutyLogo,
-    alt: "Logo of PagerDuty computer software company.",
-  },
-  {
-    url: bridgeLogo,
-    alt: "Logo of Bridge.",
-  },
-  {
-    url: weworkLogo,
-    alt: "Logo of wework.",
-  },
-  {
-    url: pcsLogo,
-    alt: "Logo of Professional Computer Support.",
-  },
-  {
-    url: uberLogo,
-    alt: "Logo of Uber Technologies Inc. company.",
-  },
-  {
-    url: codeForSFLogo,
-    alt: "Logo of Code for San Francisco.",
-  },
-  {
-    url: gsmaLogo,
-    alt: "Logo of GSMA.",
-  },
-];
